Check signs of extrema for S-number criterion
Since version 1.3 the S-number criterion checked whether the number of
extrema and zero crossings differ by at most one, following Huang et al
(1998). However, this is not sufficient to ensure that all minima in an
IMF are negative and all maxima are positive. An alternative
interpretation is that the endpoints should be counted as extrema as
well, but this leads to a too strict condition that causes some IMFs to
never converge.

Now libeemd just checks the signs directly, like was done before 1.3.
The "relaxed" S-number criterion still applies, i.e., the number of
extrema can change by one and still be considered stabilized. An endless
random oscillation between two consecutive numbers of extrema sometimes
occurs because of end effects and finite precision arithmetic.